ff8string changes
porting string fix from my fork. I just copied the files and committed to master. I've made too many changes to merge my fork to master.
